webtemplates/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
google-auth==1.31.0
requests==2.26.0
pymongo
dnspython 
PyMySQL==1.0.2
Flask==2.0.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
SQLAlchemy==1.3.8